software/gui/requirements.txt
cycler==0.10.0
joblib==0.17.0
keras==2.4.3
kiwisolver==1.0.1
matplotlib==3.0.1
mne==0.21.1
nibabel==3.2.0
numpy==1.18.5
pygame==2.0.0
pynput==1.4
pyopengl==3.1.0
pyparsing==2.3.0
pyqt5==5.11.3
pyqtgraph==0.10.0
pyserial==3.4
python-dateutil==2.7.5
python-xlib==0.23 ; 'linux' in sys_platform
qdarkstyle==2.6.1
tensorflow==2.3.1
scikit-learn==0.23.2
scipy==1.5.4
six==1.11.0
sklearn==0.0
